task3/subscribers/cloudrun_push/requirements.txt
Flask
google-cloud-pubsub
google-cloud-bigquery
google-cloud-secret-manager
gunicorn